# gui/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "all-diary API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/diaries": {
      "get": {
        "summary": "List Diaries",
        "description": "日本時間の `start_date` から `end_date` まで(両端を含む)の日記を、コメント付きで時刻の順に返す。",
        "operationId": "list_diaries_api_diaries_get",
        "parameters": [
          {
            "name": "start_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start Date"
            }
          },
          {
            "name": "end_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End Date"
            }
          },
          {
            "name": "x-diary-user",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Diary-User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DiaryRecord"
                  },
                  "title": "Response List Diaries Api Diaries Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Commit Diary",
        "operationId": "commit_diary_api_diaries_post",
        "parameters": [
          {
            "name": "x-diary-user",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Diary-User"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DiaryCreateForm"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiaryRecord"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/diaries/similar": {
      "post": {
        "summary": "Search Similar Diaries",
        "description": "`text` に似た日記の数と、似ている順の頭の何件か。",
        "operationId": "search_similar_diaries_api_diaries_similar_post",
        "parameters": [
          {
            "name": "x-diary-user",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Diary-User"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SimilarSearchRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimilarDiaries"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/on-this-day": {
      "get": {
        "summary": "List On This Day",
        "description": "`day` と同じ月日の前後 `around_days` 日の日記を、年ごとに新しい年から。",
        "operationId": "list_on_this_day_api_on_this_day_get",
        "parameters": [
          {
            "name": "day",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "Day"
            }
          },
          {
            "name": "around_days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Around Days"
            }
          },
          {
            "name": "x-diary-user",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Diary-User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OnThisDayYear"
                  },
                  "title": "Response List On This Day Api On This Day Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/diary-counts": {
      "get": {
        "summary": "Count Diaries By Day",
        "description": "日本時間の暦の日ごとの、日記の件数と文字数。書いた日だけ。",
        "operationId": "count_diaries_by_day_api_diary_counts_get",
        "parameters": [
          {
            "name": "start_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start Date"
            }
          },
          {
            "name": "end_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End Date"
            }
          },
          {
            "name": "x-diary-user",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Diary-User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DayCount"
                  },
                  "title": "Response Count Diaries By Day Api Diary Counts Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/future-diaries": {
      "get": {
        "summary": "List Future Diaries",
        "description": "まだ届いていない未来の日記の、届く時刻と書いた時刻。本文は返さない。",
        "operationId": "list_future_diaries_api_future_diaries_get",
        "parameters": [
          {
            "name": "x-diary-user",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Diary-User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FutureDiaryRecord"
                  },
                  "title": "Response List Future Diaries Api Future Diaries Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Send Future Diary",
        "operationId": "send_future_diary_api_future_diaries_post",
        "parameters": [
          {
            "name": "x-diary-user",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Diary-User"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FutureDiaryCreateForm"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FutureDiaryRecord"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/diaries/{diary_id}": {
      "get": {
        "summary": "Read Diary",
        "operationId": "read_diary_api_diaries__diary_id__get",
        "parameters": [
          {
            "name": "diary_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Diary Id"
            }
          },
          {
            "name": "x-diary-user",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Diary-User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiaryRecord"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Update Diary",
        "operationId": "update_diary_api_diaries__diary_id__patch",
        "parameters": [
          {
            "name": "diary_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Diary Id"
            }
          },
          {
            "name": "x-diary-user",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Diary-User"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DiaryUpdateForm"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiaryRecord"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Diary",
        "description": "日記を、付いたコメントごと消す。",
        "operationId": "delete_diary_api_diaries__diary_id__delete",
        "parameters": [
          {
            "name": "diary_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Diary Id"
            }
          },
          {
            "name": "x-diary-user",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Diary-User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeletedDiary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/diaries/{diary_id}/comments": {
      "get": {
        "summary": "List Comments",
        "operationId": "list_comments_api_diaries__diary_id__comments_get",
        "parameters": [
          {
            "name": "diary_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Diary Id"
            }
          },
          {
            "name": "x-diary-user",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Diary-User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CommentRecord"
                  },
                  "title": "Response List Comments Api Diaries  Diary Id  Comments Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/comments": {
      "post": {
        "summary": "Commit Comment",
        "operationId": "commit_comment_api_comments_post",
        "parameters": [
          {
            "name": "x-diary-user",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Diary-User"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommentCreateForm"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommentRecord"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/legacy-db/summary": {
      "post": {
        "summary": "Summarize Legacy Db",
        "description": "以前の SQLite の db ファイル(本文にそのまま)に、誰の日記が何件あるか。db には触れない。",
        "operationId": "summarize_legacy_db_api_legacy_db_summary_post",
        "parameters": [
          {
            "name": "x-diary-user",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Diary-User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LegacyDbSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/octet-stream": {
              "schema": {
                "type": "string",
                "format": "binary"
              }
            }
          }
        }
      }
    },
    "/api/legacy-db/import": {
      "post": {
        "summary": "Import Legacy Db",
        "description": "以前の SQLite の db ファイル(本文にそのまま)から、`source_user_id` の日記とコメントを取り込む。",
        "operationId": "import_legacy_db_api_legacy_db_import_post",
        "parameters": [
          {
            "name": "source_user_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Source User Id"
            }
          },
          {
            "name": "x-diary-user",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Diary-User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LegacyImported"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/octet-stream": {
              "schema": {
                "type": "string",
                "format": "binary"
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "DiaryCreateForm": {
        "properties": {
          "text": {
            "type": "string",
            "minLength": 1,
            "title": "Text"
          },
          "time": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Time"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "text"
        ],
        "title": "DiaryCreateForm"
      },
      "DiaryRecord": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "time": {
            "type": "string",
            "title": "Time"
          },
          "written_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Written At"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "comments": {
            "items": {
              "$ref": "#/components/schemas/CommentRecord"
            },
            "type": "array",
            "title": "Comments"
          }
        },
        "type": "object",
        "required": [
          "id",
          "time",
          "written_at",
          "text",
          "comments"
        ],
        "title": "DiaryRecord"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SimilarSearchRequest": {
        "properties": {
          "text": {
            "type": "string",
            "title": "Text"
          }
        },
        "type": "object",
        "required": [
          "text"
        ],
        "title": "SimilarSearchRequest"
      },
...
    }
  }
}